delete and cleanup older versions of files
this solves problem of what happens when several processes writing
into one file.

firstly, the table 'files' contains timestamp of the writing. It is
most a.s. simultaneous write will have a different timestamp.

a completed write will update the entry in the 'files_timestamp.

In case a file is requested to be downloaded, the current timestamp of
completely uploaded file form 'files_timestamp' is used.

A 'cleanup_old_versions' procedure can be used to clean up older
version.