fix: correct structure of the spatial index
why? a fixed length of a geohash leads to a lot of entries
corresponding to the single data entry (for example of size of geohash
box is small, and included region is big).

the new structure of the spatial index mitigates this problem by
introducing several tables for different lengths of geohash

OLD STRUCTURE:
  table 'data': data_id, data
      data_id is a hash of data
      data is a json string

  table 'hash': geohash, link
      geohash has a fixed length
      link data_id in 'data' table

NEW STRUCTURE:
  table 'data': data_id, data
      data_id is a hash of data
      data is a json string

  multiple tables 'hash#i' contains 2 columns: geohash, link
      the geohash has length '#i'
      link is either a data_id or the geohash with length '#i+1'